Refactor tracking plugin file structure
